Mute now uses modular roles

+++ commands/mute.js
@@ -1,17 +1,25 @@
 const discord = require('discord.js')
 const fs = require('fs')
-const mutedid = '495945521408770049'
 
-function muteUser (client, member, duration, muter, channel) {
-  var roles = member.roles
-  var roleids = []
-  roles.forEach(function (role) {
-    roleids.push(role.id)
-  })
+function muteUser(client, member, duration, muter, channel) {
+  let roles = member.roles
+  let roleIDs = roles.map(role => role.id)
 
+  // Get the muted role from guild role info
+  const roleData = client.serverRoles.get(channel.guild.id)
+  if (!roleData) return
+  if (!roleData.muted) return
+  const muteID = roleData.muted
+
+  // Create the locally stored mutes storage if it doesn't exist
+  if (client.mutesData == null) {
+    client.mutesData = new discord.Collection()
+  }
+
+  // Setup the options for the mute
   var options = {
     guild: member.guild.id,
-    roles: roleids,
+    roles: roleIDs,
     unmute: new Date(Date.now() + duration * 60000),
     muter: muter.id,
     channel: channel.id
@@ -19,26 +27,35 @@ function muteUser (client, member, duration, muter, channel) {
   client.mutesData.set(member.id, options)
 
   // Remove all roles, then add muted role
+  // We need this forEach loop vs removeRoles in the case of un-removable roles
+  // eg. Nitro Boost
   roles.forEach(function (role) {
-    if (role.id === mutedid) return
-    member.removeRole(role).catch(function (e) {})
+    if (role.id === muteID) return
+    member.removeRole(role).catch(function (e) { })
   })
-  member.addRole(member.guild.roles.get(mutedid))
+  member.addRole(member.guild.roles.get(muteID))
 
   // Write a data file in case of restarting
   client.writeDataFile('mutes', member.id, options)
 }
 
-function unmuteUser (client, id) {
-  var settings = client.mutesData.get(id)
-  var guild = client.guilds.get(settings.guild)
-  var member = guild.members.get(id)
+function unmuteUser(client, id) {
+  const settings = client.mutesData.get(id)
+  const guild = client.guilds.get(settings.guild)
+  const member = guild.members.get(id)
+  const channel = guild.channels.get(settings.channel)
   client.mutesData.delete(id)
+
+  // Get the muted role from guild role info
+  const roleData = client.serverRoles.get(channel.guild.id)
+  if (!roleData) return
+  if (!roleData.muted) return
+  const muteID = roleData.muted
 
   // Delete the mute data file (if it exists)
   try {
-    fs.unlink('data/mutes/' + member.id + '.json', function (e) {})
-  } catch (e) {}
+    fs.unlink('data/mutes/' + member.id + '.json', function (e) { })
+  } catch (e) { }
 
   // Abort if the member doesn't exist
   if (!member) {
@@ -47,16 +64,15 @@ function unmuteUser (client, id) {
 
   // Add roles back, then removed muted role
   settings.roles.forEach(function (id) {
-    var role = guild.roles.get(id)
-    member.addRole(role).catch(function (e) {})
+    const role = guild.roles.get(id)
+    member.addRole(role).catch(function (e) { })
   })
-  member.removeRole(member.guild.roles.get(mutedid))
+  member.removeRole(member.guild.roles.get(muteID))
 
   // Reply message
   try {
-    var channel = guild.channels.get(settings.channel)
     channel.send(member.displayName + ' has been unmuted')
-  } catch (e) {}
+  } catch (e) { }
 }
 
 module.exports = {
@@ -65,18 +81,12 @@ module.exports = {
   aliases: ['banish', 'void'],
   mute: muteUser,
   unmute: unmuteUser,
-  execute (message, args, client) {
-    if (message.guild.id !== '309951255575265280') return
-
+  guilds: ['309951255575265280'],
+  execute(message, args, client) {
     // Check that the user has permission
     if (!client.isModerator(message.member)) {
       message.channel.send('You need to be a Moderator to use this!')
       return
-    }
-
-    // Create the locally stored mutes storage if it doesn't exist
-    if (client.mutesData == null) {
-      client.mutesData = new discord.Collection()
     }
 
     try {
